DOC/AuditoriaHospitalar/requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
pandas
numpy
google.generativeai
chromadb
sentence-transformers
python-dotenv

